Use the same EidasLevel enum
